+++ NEWS.md
@@ -1,1 +1,7 @@
 # rvest 0.1.0.9000
+
+* Add xml support: parse with `xml()`, then work with using `xml_node()`,
+  `xml_attr()`, `xml_attrs()`, `xml_text()` and `xml_tag()` (#24).
+
+* `xml_structure()`: new function that displays the structure (i.e. tag
+  and attribute names) of a xml/html object (#10).